+++ src/components/CardList.vue
@@ -7,14 +7,10 @@ defineProps({
 })
 
 const emit = defineEmits(['addToFavorite', 'addToCart'])
-
-const onClickAdd = () => {
-  alert('Добавить')
-}
 </script>
 
 <template>
-  <div class="grid grid-cols-4 gap-5">
+  <div class="grid grid-cols-4 gap-5" v-auto-animate>
     <Card
       v-for="item in items"
       :key="item.id"
