+++ src/components/CartItemList.vue
@@ -6,7 +6,7 @@ const { cart, removeFromCart } = inject('cart')
 </script>
 
 <template>
-  <div class="flex flex-col gap-4">
+  <div class="flex flex-col gap-4 justify-between" v-auto-animate>
     <CartItem
       v-for="item in cart"
       :key="item.id"
