+++ src/main.js
@@ -1,6 +1,11 @@
 import './assets/main.css'
 
 import { createApp } from 'vue'
+import { autoAnimatePlugin } from '@formkit/auto-animate/vue'
 import App from './App.vue'
 
-createApp(App).mount('#app')
+const app = createApp(App)
+
+app.use(autoAnimatePlugin)
+
+app.mount('#app')
