+++ src/pages/home.vue
@@ -1,0 +1,82 @@
+<script setup>
+import { inject } from 'vue'
+import CardList from '../components/CardList.vue'
+import axios from 'axios'
+
+const { addToCart, removeFromCart } = inject('cart')
+
+const filters = reactive({
+  sortBy: 'title',
+  searchQuery: ''
+})
+
+const onClickAddPlus = (item) => {
+  if (!item.isAdded) {
+    addToCart(item)
+  } else {
+    removeFromCart(item)
+  }
+  console.log(cart)
+}
+
+const onChangeSelect = (event) => {
+  filters.sortBy = event.target.value
+}
+
+const onChangeSearchInput = (event) => {
+  filters.searchQuery = event.target.value
+}
+
+const addToFavorite = async (item) => {
+  try {
+    if (!item.isFavorite) {
+      const obj = {
+        parentId: item.id
+      }
+
+      item.isFavorite = true
+
+      const { data } = await axios.post('https://c0abc61e8e424c50.mokky.dev/favorites', obj)
+
+      item.favoriteId = data.id
+
+      console.log(data)
+    } else {
+      item.isFavorite = false
+
+      await axios.delete(`https://c0abc61e8e424c50.mokky.dev/favorites/${item.favoriteId}`)
+
+      item.favoriteId = null
+    }
+  } catch (err) {
+    console.log(err)
+  }
+}
+</script>
+
+<template>
+  <div class="flex justify-between items-center">
+    <h2 class="text-3xl font-bold mb-8">Все кроссовки</h2>
+
+    <div class="flex gap-4">
+      <select @change="onChangeSelect" class="py-2 px-3 border rounded-md outline-none">
+        <option value="name">По названию</option>
+        <option value="price">По цене (дешевые)</option>
+        <option value="-price">По цене (дорогие)</option>
+      </select>
+
+      <div class="relative">
+        <img class="absolute left-4 top-3" src="/search.svg" />
+        <input
+          @input="onChangeSearchInput"
+          class="border rounded-md py-2 pl-11 pr-4 outline-none focus:border-gray-400"
+          placeholder="Поиск..."
+        />
+      </div>
+    </div>
+  </div>
+
+  <div class="mt-10">
+    <CardList :items="items" @add-to-favorite="addToFavorite" @add-to-cart="onClickAddPlus" />
+  </div>
+</template>
